orchestrator/gui/server/requirements.txt
bleach
channels
cherrypy
django
django-cors-headers
djangorestframework
djangorestframework-datatables
djangorestframework-jwt
django-rest-swagger
mysql-connector-python